pyproject.toml
[tool.poetry]
name = "testbot3"
version = "0.1.0"
description = ""
authors = ["Mihir <mihirniyogi942@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
python-telegram-bot = "^13.14"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
